Refactored SQS integration into its own module and refactored tests
